api/requirements.txt
Django>=4.1
psycopg2>=2.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.2 
pandas==1.5.2
